Refactored the code that extracts the version of ROS and ubuntu distribution
This helps keep similar logic used for discovery of ROS and ubuntu
distribution version in the same place. This will make it easier to
update in future.
