+++ colcon_bundle/installer/apt.py
@@ -4,13 +4,13 @@
 import glob
 import json
 import os
-import platform
 import subprocess
 
 from apt.cache import FetchFailedException
 from apt.package import FetchError
 from colcon_bundle.installer import BundleInstallerExtensionPoint
 from colcon_bundle.verb import logger
+from colcon_bundle.verb.utilities import get_ubuntu_distribution_version
 from colcon_core.plugin_system import satisfies_version
 
 
@@ -38,10 +38,8 @@ class AptBundleInstallerExtension(BundleInstallerExtensionPoint):
             assets_directory, 'apt_package_blacklist.txt')
         sources_list_path = os.path.join(assets_directory, 'xenial.sources.list')
 
-        if platform.linux_distribution:
-            distribution = platform.linux_distribution()
-            if len(distribution) >= 3 and distribution[2] == 'bionic':
-                sources_list_path = os.path.join(assets_directory, 'bionic.sources.list')
+        if get_ubuntu_distribution_version() == 'bionic':
+            sources_list_path = os.path.join(assets_directory, 'bionic.sources.list')
 
         parser.add_argument(
             '--apt-package-blacklist', default=blacklist_path,
