+++ colcon_bundle/verb/utilities.py
@@ -13,6 +13,21 @@ import sys
 from colcon_bundle.verb import logger
 
 
+def get_ros_distribution_version():
+    ros_distribution_version = 'kinetic'
+    if get_ubuntu_distribution_version() == 'bionic':
+        ros_distribution_version = 'melodic'
+    return
+
+def get_ubuntu_distribution_version():
+    ubuntu_distribution_version = 'xenial'
+    if platform.linux_distribution:
+        distribution = platform.linux_distribution()
+        if len(distribution) >= 3:
+            ubuntu_distribution_version = distribution[2]
+
+    return ubuntu_distribution_version
+
 def update_shebang(path):
     """
     Search for python shebangs in path and all sub-paths.
@@ -133,14 +148,9 @@ def rewrite_catkin_package_path(base_path):
     python_regex = re.compile('/usr/bin/python')
     logger.info('Starting shebang update...')
 
-    ros_distro = 'kinetic'
-    if platform.linux_distribution:
-        distribution = platform.linux_distribution()
-        if len(distribution) >= 3 and distribution[2] == 'bionic':
-            ros_distro = 'melodic'
-    
+    ros_distribution_version = get_ros_distribution_version()
     profiled_path = os.path.join(
-        base_path, 'opt', 'ros', ros_distro, 'etc', 'catkin', 'profile.d',
+        base_path, 'opt', 'ros', ros_distribution_version, 'etc', 'catkin', 'profile.d',
         '1.ros_package_path.sh')
     if os.path.isfile(profiled_path):
         with open(profiled_path, 'rb+') as file_handle:
